fix second import bug

+++ src/App.js
@@ -312,13 +312,12 @@ class App extends Component {
     var searchParams = new URLSearchParams(window.location.search);
     if (searchParams.has("import")) {
       const importPortfolio = JSON.parse(atob(searchParams.get("import")));
-// failing even if new import :(
-//       const alreadyImported = importPortfolio == localStorage.getItem("lastImport");
+      const alreadyImported = searchParams.get("import") == localStorage.getItem("lastImport");
 
-//       if (alreadyImported) {
-//         console.log('already imported this portfolio');
-//         window.location.search = "";
-//       } else {
+      if (alreadyImported) {
+        console.log('already imported this portfolio');
+        window.location.search = "x";
+      } else {
         // import the new portfolio
         if (importPortfolio.pref) {
           localStorage.setItem("pref", importPortfolio.pref);
@@ -326,9 +325,9 @@ class App extends Component {
         if (importPortfolio.coinz) {
           localStorage.setItem("coinz", importPortfolio.coinz);
         }
-        localStorage.setItem("lastImport", importPortfolio);
+        localStorage.setItem("lastImport", searchParams.get("import"));
         window.location.search = ""
-//       }
+      }
     }
